Made some minor changes to 4_16.c file

+++ WS3/4_16.c
@@ -27,6 +27,6 @@ int main(int argc, char *argv[])
         }
     }
     double avg = total/count;
-    printf("%d %d\n", max, (int)avg);
+    printf("%d %.2lf\n", max, avg);
     return 0;
 }
